requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile '.\requirements.in'
#
black==22.3.0
    # via -r .\requirements.in
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via pynacl
charset-normalizer==2.0.12
    # via requests
click==8.1.2
    # via black
colorama==0.4.4
    # via click
deprecated==1.2.13
    # via pygithub
idna==3.3
    # via requests
mypy-extensions==0.4.3
    # via black
pathspec==0.9.0
    # via black
platformdirs==2.5.1
    # via black
pycparser==2.21
    # via cffi
pygithub==1.55
    # via -r .\requirements.in
pyjwt==2.3.0
    # via pygithub
pynacl==1.5.0
    # via pygithub
python-dotenv==0.20.0
    # via -r .\requirements.in
requests==2.27.1
    # via pygithub
tomli==2.0.1
    # via black
typing-extensions==4.1.1
    # via black
urllib3==1.26.9
    # via requests
wrapt==1.14.0
    # via deprecated
